MetaGPT-EUTariffSim/requirements.txt
# 核心依赖
metagpt>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM相关
openai>=1.0.0
anthropic>=0.7.0
zhipuai>=2.0.0
qianfan>=0.3.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 网络分析
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 配置管理
pyyaml>=6.0
toml>=0.10.0
python-dotenv>=1.0.0

# 测试与验证
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 其他工具
tqdm>=4.65.0
loguru>=0.7.0
rich>=13.5.0

